Move formatter to ValidatedData
