--- radix-engine/src/utils/mod.rs
@@ -1,3 +1,0 @@
-mod formatter;
-
-pub use formatter::{format_data, format_data_with_ledger};
